fefactore code - interfaces and handler
