+++ src/interfaces/actionMouse/drag/index.ts
@@ -1,0 +1,3 @@
+import { IGetMouseMove } from '../move'
+
+export type IGetMouseDrag = IGetMouseMove
