+++ src/interfaces/actionMouse/move/index.ts
@@ -1,0 +1,5 @@
+import { Point } from '@nut-tree/nut-js'
+
+export interface IGetMouseMove {
+  (direction: Point[] | Promise<Point[]>): void;
+}
