+++ src/interfaces/index.ts
@@ -1,0 +1,1 @@
+// export { IGetMouseDrag, IGetMouseMove, IActionMouseClick, IHandlerWebServer, TActionMouseClick } from './actionMouse'
